Feature traffic light predicate

See merge request maierhofer/environment-model!13

+++ src/commonroad_cpp/auxiliaryDefs/structs.h
@@ -10,7 +10,7 @@
 
 struct TrafficLightCycleElement {
     TrafficLightState color;
-    int duration;
+    size_t duration;
 };
 
 struct vertex {
